requirements.txt
fastapi
uvicorn
opencv-python
pillow
tensorflow==2.12.0
numpy
gdown
